Proof of concept
Another approach, without extra variables. Makes use of the 'hidden' tag for a button. Please have a look.

Also added new icons in style.css to replace the < on the button which gives a warning in the validator and has to be escaped, making the code in a button almost unreadable.