Try to remove DisplayInterval
it it works we should simplify Datum_Dag, Datum_Dag AS datumtijd
as well

For me it looks like this parameter is not needed, maybe from very old day with flash chart
but with highchart not needed,
We show the data which is in the database, there it depends if you write data per minute, 5min, or 1h.
Highchart should be able to handle it. If it is too much data, -> change the Importer to only write data per x/min

If you have checked this as well, I'll remove parameter and optimize SQL