New approach to find Max day
Reworked the query to sum all inverter values to get the date with best kwh. Let Highcharts calculate the max values. But in order to work consistent, load order for series is changed. When loaded series get an index. This gives some additional challenges. WIP for some small things like the actual power generated (which is not correct atm, but I need solar energy to work on this).